// src/features/design-options/types/design-option.ts
// First we define the model
export type DesignOption = {
  id: string;

  // Core info
  name: string;
  description: string;
  area: string;

  // Comparison metrics
  embodiedCarbon: number;
  daylightScore: number;
  costEstimate: number;
  programFit: number;

  // Extra info
  tags: string[];
  notes?: string;

  // Metadata
  source: 'manual' | 'imported';
  // createdAT and updatedAt are automatically generated by supabase
  createdAt: string;
  updatedAt: string;

  // It adds extra value:
  pros: string[]; //detail view
  cons: string[]; //detail view
  referenceUrl?: string[]; //External URL as a reference
};